+++ src/vm/errors/vm_errors.rs
@@ -1,4 +1,5 @@
 use super::exec_scope_errors::ExecScopeError;
+use super::hint_errors::HintError;
 use super::trace_errors::TraceError;
 use crate::types::relocatable::{MaybeRelocatable, Relocatable};
 use crate::vm::errors::memory_errors::MemoryError;
@@ -28,6 +29,8 @@ pub enum VirtualMachineError {
     UnconstrainedResJumpRel,
     #[error("Res.UNCONSTRAINED cannot be used with Opcode.ASSERT_EQ")]
     UnconstrainedResAssertEq,
+    #[error("Couldn't compute operand {0} at address {1}")]
+    FailedToComputeOperands(String, Relocatable),
     #[error("An ASSERT_EQ instruction failed: {0} != {1}.")]
     DiffAssertValues(MaybeRelocatable, MaybeRelocatable),
     #[error("Call failed to write return-pc (inconsistent op0): {0} != {1}. Did you forget to increment ap?")]
@@ -68,32 +71,24 @@ pub enum VirtualMachineError {
     ExpectedInteger(MaybeRelocatable),
     #[error("Expected relocatable at address {0}")]
     ExpectedRelocatable(MaybeRelocatable),
-    #[error("Failed to get ids for hint execution")]
-    FailedToGetIds,
-    #[error("Assertion failed, {0}, is not less or equal to {1}")]
-    NonLeFelt(BigInt, BigInt),
-    #[error("Div out of range: 0 < {0} <= {1}")]
-    OutOfValidRange(BigInt, BigInt),
-    #[error("Assertion failed, 0 <= ids.a % PRIME < range_check_builtin.bound \n a = {0} is out of range")]
-    ValueOutOfRange(BigInt),
     #[error("Value: {0} should be positive")]
     ValueNotPositive(BigInt),
-    #[error("Unknown Hint: {0}")]
-    UnknownHint(String),
-    #[error("Value: {0} is outside valid range")]
-    ValueOutsideValidRange(BigInt),
-    #[error("split_int(): value is out of range")]
-    SplitIntNotZero,
-    #[error("split_int(): Limb {0} is out of range.")]
-    SplitIntLimbOutOfRange(BigInt),
+    #[error("Div out of range: 0 < {0} <= {1}")]
+    OutOfValidRange(BigInt, BigInt),
     #[error("Failed to compare {0} and {1}, cant compare a relocatable to an integer value")]
     DiffTypeComparison(MaybeRelocatable, MaybeRelocatable),
-    #[error("assert_not_equal failed: {0} =  {1}")]
-    AssertNotEqualFail(MaybeRelocatable, MaybeRelocatable),
     #[error("Failed to compare {0} and  {1}, cant compare two relocatable values of different segment indexes")]
     DiffIndexComp(Relocatable, Relocatable),
-    #[error("Value: {0} is outside of the range [0, 2**250)")]
-    ValueOutside250BitRange(BigInt),
+    #[error("Couldn't convert BigInt to usize")]
+    BigintToUsizeFail,
+    #[error("Couldn't convert BigInt to u64")]
+    BigintToU64Fail,
+    #[error("Couldn't convert BigInt to u32")]
+    BigintToU32Fail,
+    #[error("Couldn't convert usize to u32")]
+    UsizeToU32Fail,
+    #[error("None value was found in memory range cell")]
+    NoneInMemoryRange,
     #[error("Can't calculate the square root of negative number: {0})")]
     SqrtNegative(BigInt),
     #[error("{0} is not divisible by {1}")]
@@ -104,136 +99,22 @@ pub enum VirtualMachineError {
     DividedByZero,
     #[error("Failed to calculate the square root of: {0})")]
     FailedToGetSqrt(BigInt),
-    #[error("Assertion failed, {0} % {1} is equal to 0")]
-    AssertNotZero(BigInt, BigInt),
-    #[error(transparent)]
-    MainScopeError(#[from] ExecScopeError),
-    #[error("Failed to get scope variables")]
-    ScopeError,
-    #[error("Variable {0} not present in current execution scope")]
-    VariableNotInScopeError(String),
-    #[error("DictManagerError: Tried to create tracker for a dictionary on segment: {0} when there is already a tracker for a dictionary on this segment")]
-    CantCreateDictionaryOnTakenSegment(isize),
-    #[error("Dict Error: No dict tracker found for segment {0}")]
-    NoDictTracker(isize),
-    #[error("ict Error: No value found for key: {0}")]
-    NoValueForKey(BigInt),
-    #[error("Assertion failed, a = {0} % PRIME is not less than b = {1} % PRIME")]
-    AssertLtFelt(BigInt, BigInt),
-    #[error("find_element() can only be used with n_elms <= {0}.\nGot: n_elms = {1}")]
-    FindElemMaxSize(BigInt, BigInt),
-    #[error(
-        "Invalid index found in find_element_index. Index: {0}.\nExpected key: {1}, found_key {2}"
-    )]
-    InvalidIndex(BigInt, BigInt, BigInt),
-    #[error("Found Key is None")]
-    KeyNotFound,
-    #[error("AP tracking data is None; could not apply correction to address")]
-    NoneApTrackingData,
-    #[error("Tracking groups should be the same, got {0} and {1}")]
-    InvalidTrackingGroup(usize, usize),
-    #[error("Expected relocatable for ap, got {0}")]
-    InvalidApValue(MaybeRelocatable),
-    #[error("Dict Error: Tried to create a dict whithout an initial dict")]
-    NoInitialDict,
-    #[error("squash_dict_inner fail: couldnt find key {0} in accesses_indices")]
-    NoKeyInAccessIndices(BigInt),
-    #[error("squash_dict_inner fail: local accessed_indices is empty")]
-    EmptyAccessIndices,
-    #[error("squash_dict_inner fail: local current_accessed_indices is empty")]
-    EmptyCurrentAccessIndices,
-    #[error("squash_dict_inner fail: local current_accessed_indices not empty, loop ended with remaining unaccounted elements")]
-    CurrentAccessIndicesNotEmpty,
-    #[error("Dict Error: Got the wrong value for dict_update, expected value: {0}, got: {1} for key: {2}")]
-    WrongPrevValue(BigInt, BigInt, BigInt),
-    #[error("squash_dict_inner fail: Number of used accesses:{0} doesnt match the lengh: {1} of the access_indices at key: {2}")]
-    NumUsedAccessesAssertFail(BigInt, usize, BigInt),
-    #[error("squash_dict_inner fail: local keys is not empty")]
-    KeysNotEmpty,
-    #[error("squash_dict_inner fail: No keys left but remaining_accesses > 0")]
-    EmptyKeys,
-    #[error("squash_dict fail: Accesses array size must be divisible by DictAccess.SIZE")]
-    PtrDiffNotDivisibleByDictAccessSize,
-    #[error("squash_dict() can only be used with n_accesses<={0}. ' \nGot: n_accesses={1}")]
-    SquashDictMaxSizeExceeded(BigInt, BigInt),
-    #[error("squash_dict fail: n_accesses: {0} is too big to be converted into an iterator")]
-    NAccessesTooBig(BigInt),
-    #[error("Couldn't convert BigInt to usize")]
-    BigintToUsizeFail,
-    #[error("Couldn't convert BigInt to u64")]
-    BigintToU64Fail,
-    #[error("Couldn't convert BigInt to u32")]
-    BigintToU32Fail,
-    #[error("Couldn't convert usize to u32")]
-    UsizeToU32Fail,
-    #[error("usort() can only be used with input_len<={0}. Got: input_len={1}.")]
-    UsortOutOfRange(u64, BigInt),
-    #[error("unexpected usort fail: positions_dict or key value pair not found")]
-    UnexpectedPositionsDictFail,
-    #[error("unexpected verify multiplicity fail: positions not found")]
-    PositionsNotFound,
-    #[error("unexpected verify multiplicity fail: positions length != 0")]
-    PositionsLengthNotZero,
-    #[error("unexpected verify multiplicity fail: couldn't pop positions")]
-    CouldntPopPositions,
-    #[error("unexpected verify multiplicity fail: last_pos not found")]
-    LastPosNotFound,
-    #[error("Set's starting point {0} is bigger it's ending point {1}")]
-    InvalidSetRange(MaybeRelocatable, MaybeRelocatable),
-    #[error("Failed to construct a fixed size array of size: {0}")]
-    FixedSizeArrayFail(usize),
-    #[error("{0}")]
-    AssertionFailed(String),
-    #[error("Wrong dict pointer supplied. Got {0}, expected {1}.")]
-    MismatchedDictPtr(Relocatable, Relocatable),
-    #[error("Integer must be postive or zero, got: {0}")]
-    SecpSplitNegative(BigInt),
-    #[error("Integer: {0} out of range")]
-    SecpSplitutOfRange(BigInt),
-    #[error("verify_zero: Invalid input {0}")]
-    SecpVerifyZero(BigInt),
-    #[error("Cant substract {0} from offset {1}, offsets cant be negative")]
-    CantSubOffset(usize, usize),
-    #[error("unsafe_keccak() can only be used with length<={0}. Got: length={1}")]
-    KeccakMaxSize(BigInt, BigInt),
-    #[error("Invalid word size: {0}")]
-    InvalidWordSize(BigInt),
-    #[error("Invalid input length, Got: length={0}")]
-    InvalidKeccakInputLength(BigInt),
-    #[error("None value was found in memory range cell")]
-    NoneInMemoryRange,
     #[error("Expected integer, found: {0:?}")]
     ExpectedIntAtRange(Option<MaybeRelocatable>),
-    #[error("Expected size to be in the range from [0, 100), got: {0}")]
-    InvalidKeccakStateSizeFelts(BigInt),
-    #[error("Expected size to be in range from [0, 10), got: {0}")]
-    InvalidBlockSize(BigInt),
     #[error("Could not convert slice to array")]
     SliceToArrayError,
-    #[error("HintProcessor failed retrieve the compiled data necessary for hint execution")]
-    WrongHintData,
     #[error("Failed to compile hint: {0}")]
     CompileHintFail(String),
     #[error("op1_addr is Op1Addr.IMM, but no immediate was given")]
     NoImm,
-    #[error("Tried to compute an address but there was no register in the reference.")]
-    NoRegisterInReference,
-    #[error("Couldn't compute operand {0} at address {1}")]
-    FailedToComputeOperands(String, Relocatable),
-    #[error("Custom Hint Error: {0}")]
-    CustomHint(String),
-    #[error("Arc too big, {0} must be <= {1} and {2} <= {3}")]
-    ArcTooBig(BigInt, BigInt, BigInt, BigInt),
-    #[error("Excluded is supposed to be 2, got {0}")]
-    ExcludedNot2(BigInt),
+    #[error("Cant substract {0} from offset {1}, offsets cant be negative")]
+    CantSubOffset(usize, usize),
     #[error("Execution reached the end of the program. Requested remaining steps: {0}.")]
     EndOfProgram(usize),
-    #[error("Missing constant: {0}")]
-    MissingConstant(&'static str),
-    #[error("Fail to get constants for hint execution")]
-    FailedToGetConstant,
     #[error(transparent)]
     TracerError(#[from] TraceError),
+    #[error(transparent)]
+    MainScopeError(#[from] ExecScopeError),
     #[error("Current run is not finished")]
     RunNotFinished,
     #[error("Invalid argument count, expected {0} but got {1}")]
@@ -241,5 +122,7 @@ pub enum VirtualMachineError {
     #[error("{0}, {1}")]
     ErrorMessageAttribute(String, Box<VirtualMachineError>),
     #[error("Got an exception while executing a hint: {1}")]
-    Hint(usize, Box<VirtualMachineError>),
+    Hint(usize, Box<HintError>),
+    #[error("Unexpected Failure")]
+    Unexpected,
 }