+++ CHANGELOG.md
@@ -22,6 +22,11 @@
     * Internal changes: 
         * Make the `compile_hint` implementation which was in the `BuiltinHintProcessor` the default implementation in the trait.
 
+* Add new error type `HintError` [#676](https://github.com/lambdaclass/cairo-rs/pull/676)
+    * Public Api changes:
+        * `HintProcessor::execute_hint()` now returns a `HintError` instead of a `VirtualMachineError`
+        * helper functions on `hint_processor_utils.rs` now return a `HintError`
+
 * Use CairoArg enum instead of Any in CairoRunner::run_from_entrypoint [#686](https://github.com/lambdaclass/cairo-rs/pull/686)
     * Public Api changes:
         * Remove `Result` from `MaybeRelocatable::mod_floor`, it now returns a `MaybeRelocatable`